requirements.txt
interactions.py
python-dotenv
mysql-connector-python
